Update Substitution Cipher Encryption With Key
This Substitution Cipher With Key takes an entered text and encrypts it for the user and prints the resulting text. It uses a specific key unique to this program for encrypting and decrypting messages. It takes spaces and punctuation and symbols and prints them in the right places. It only performs the encryption on the text entered and not the spaces/punctuation/symbols.